# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ChillBooking API",
        "description": "API documentation for ChillBooking service",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get all bookings",
                "operationId": "2f7e7260c708227e4c44de9d5b59ea36",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingResource"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unable to fetch bookings"
                    }
                }
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "9ec8aa3de0d1efdd0cfe72c27eac745c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unable to create booking"
                    }
                }
            }
        },
        "/api/v1/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get booking by ID",
                "operationId": "fd0155f99c405cf6978c540b2fd76f64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Update a booking",
                "operationId": "c05e14803d319fd8e11c1e58dfeef2f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unable to update booking"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Delete a booking",
                "operationId": "cac928061de061c77bfc403f57876a18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking deleted successfully"
                    },
                    "500": {
                        "description": "Unable to delete booking"
                    }
                }
            }
        },
        "/api/v1/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Retrieve list of locations",
                "operationId": "49873ce3b7837cc865db7849ec7cccda",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LocationResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Create a new location",
                "operationId": "aff6a55f233e5e2483ad2a6b68576fbb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LocationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Location created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/locations/{id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Retrieve a specific location",
                "operationId": "297bb90722857403e546b74145806a12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Locations"
                ],
                "summary": "Update a specific location",
                "operationId": "f8115ec393e6a83bc212ee2a6bbb5ee7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LocationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LocationResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Delete a specific location",
                "operationId": "732da05f06d706d822b05dd8c62945aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Location deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BookingRequest": {
                "title": "Booking Request",
                "description": "Booking request payload",
                "required": [
                    "booking_status",
                    "payment_status"
                ],
                "properties": {
                    "booking_status": {
                        "description": "Status of the booking",
                        "type": "integer"
                    },
                    "payment_status": {
                        "description": "Payment status of the booking",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the customer",
                        "type": "string"
                    },
                    "company": {
                        "description": "Company of the customer",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email of the customer",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "Phone number of the customer",
                        "type": "string"
                    },
                    "job_reference": {
                        "description": "Job reference for the booking",
                        "type": "string"
                    },
                    "dangerous_goods": {
                        "description": "Whether the booking involves dangerous goods",
                        "type": "boolean"
                    },
                    "pickup_detail": {
                        "description": "Pickup detail",
                        "type": "object"
                    },
                    "delivery_detail": {
                        "description": "Delivery detail",
                        "type": "object"
                    },
                    "final_price": {
                        "description": "Final price of the booking",
                        "type": "number",
                        "format": "float"
                    },
                    "total_qty": {
                        "description": "Total quantity of items",
                        "type": "integer"
                    },
                    "total_spaces": {
                        "description": "Total number of spaces",
                        "type": "integer"
                    },
                    "total_volume": {
                        "description": "Total volume of the booking",
                        "type": "integer"
                    },
                    "total_weight": {
                        "description": "Total weight of the booking",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "LocationRequest": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "description": "Title of the location",
                        "type": "string",
                        "maxLength": 255
                    },
                    "lng": {
                        "description": "Longitude of the location",
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "lat": {
                        "description": "Latitude of the location",
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "address": {
                        "description": "Address of the location",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "city": {
                        "description": "City of the location",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "postcode": {
                        "description": "Postcode of the location",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "BookingResource": {
                "title": "Booking Resource",
                "description": "Booking resource representation",
                "properties": {
                    "id": {
                        "description": "ID of the booking",
                        "type": "integer"
                    },
                    "booking_status": {
                        "description": "Status of the booking",
                        "type": "integer"
                    },
                    "payment_status": {
                        "description": "Payment status of the booking",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the customer",
                        "type": "string"
                    },
                    "company": {
                        "description": "Company of the customer",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email of the customer",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "Phone number of the customer",
                        "type": "string"
                    },
                    "job_reference": {
                        "description": "Job reference for the booking",
                        "type": "string"
                    },
                    "dangerous_goods": {
                        "description": "Whether the booking involves dangerous goods",
                        "type": "boolean"
                    },
                    "pickup_detail_id": {
                        "description": "Pickup detail ID",
                        "type": "integer"
                    },
                    "delivery_detail_id": {
                        "description": "Delivery detail ID",
                        "type": "integer"
                    },
                    "final_price": {
                        "description": "Final price of the booking",
                        "type": "number",
                        "format": "float"
                    },
                    "total_qty": {
                        "description": "Total quantity of items",
                        "type": "integer"
                    },
                    "total_spaces": {
                        "description": "Total number of spaces",
                        "type": "integer"
                    },
                    "total_volume": {
                        "description": "Total volume of the booking",
                        "type": "integer"
                    },
                    "total_weight": {
                        "description": "Total weight of the booking",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "LocationResource": {
                "properties": {
                    "id": {
                        "description": "ID of the location",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the location",
                        "type": "string"
                    },
                    "lng": {
                        "description": "Longitude of the location",
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "lat": {
                        "description": "Latitude of the location",
                        "type": "number",
                        "format": "double",
                        "nullable": true
                    },
                    "address": {
                        "description": "Address of the location",
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "description": "City of the location",
                        "type": "string",
                        "nullable": true
                    },
                    "postcode": {
                        "description": "Postcode of the location",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "schedules": {
                        "description": "List of schedules for the location",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LocationScheduleResource"
                        }
                    }
                },
                "type": "object"
            },
            "LocationScheduleResource": {
                "properties": {
                    "id": {
                        "description": "ID of the schedule",
                        "type": "integer"
                    },
                    "location_id": {
                        "description": "ID of the location",
                        "type": "integer"
                    },
                    "day_number": {
                        "description": "Day number (0-6)",
                        "type": "integer"
                    },
                    "from": {
                        "description": "Opening time",
                        "type": "string",
                        "format": "time",
                        "nullable": true
                    },
                    "to": {
                        "description": "Closing time",
                        "type": "string",
                        "format": "time",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "location": {
                        "$ref": "#/components/schemas/LocationResource"
                    }
                },
                "type": "object"
            }
        }
    },
...
}